Adding the scripts for the Biofeedback Project
This folder contains the GUI needed for the Biofeedback Experiment.
All the necessary functions and the files for running the GUI is also
included in the folder, making this a independent folder which can run
on itself.
Only thing which need to be added to the path is the EEGLab